% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQI.R
\name{AQI_Daily}
\alias{AQI_Daily}
\title{Calculate the daily AQI score and the featured pollutants based on 6 pollutants' concentration}
\usage{
AQI_Daily(SO2, NO2, CO, PM10, PM2.5, O3)
}
\arguments{
\item{SO2}{daily concentration of \code{SO2}, μg/m3}

\item{NO2}{daily concentration of \code{NO2}, μg/m3}

\item{CO}{daily concentration of \code{CO}, mg/m3}

\item{PM10}{daily concentration of \code{PM10}, μg/m3}

\item{PM2.5}{daily concentration of \code{PM2.5}, μg/m3}

\item{O3}{daily concentration of \code{O3}, μg/m3}
}
\value{
a named number of the daily AQI score, the name is the featured pollutant
}
\description{
Calculate the daily AQI score and the featured pollutants based on 6 pollutants' concentration
}
\examples{
Calc_Daily_AQI(SO2 = 55, NO2 = 23, CO = 12, O3 = 122, PM2.5 = 35, PM10 = 55)
}
