% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQI.R
\name{Calc_Hourly_AQI}
\alias{Calc_Hourly_AQI}
\title{Calculate the hourly AQI score and the featured pollutants based on 4 pollutants' concentration}
\usage{
Calc_Hourly_AQI(SO2, NO2, CO, O3)
}
\arguments{
\item{SO2}{hourly concentration of \code{SO2}, μg/m3}

\item{NO2}{hourly concentration of \code{NO2}, μg/m3}

\item{CO}{hourly concentration of \code{CO}, mg/m3}

\item{O3}{hourly concentration of \code{O3}, μg/m3}
}
\value{
a named number of the hourly AQI score, the name is the featured pollutant
}
\description{
Calculate the hourly AQI score and the featured pollutants based on 4 pollutants' concentration
}
\examples{
Calc_Hourly_AQI(SO2 = 55, NO2 = 23, CO = 140, O3 = 122)
}
