% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQI.R
\name{DaliyMeanConc}
\alias{DaliyMeanConc}
\title{Calculate daily mean concentration of 6 pollutants}
\usage{
DaliyMeanConc(Pollu, Conc)
}
\arguments{
\item{Pollu}{name of pollutants now supports \code{SO2}, \code{NO2}, \code{CO}, \code{PM2.5}, \code{PM10} and \code{O3}}

\item{Conc}{a 24-long vector of hourly average concentrations}
}
\value{
a number of daily mean concentration of 6 pollutants in which \code{O3} is calculated by maximum 8-hour average
}
\description{
Calculate daily mean concentration of 6 pollutants
}
\examples{
C <- rnorm(24, 35, 5)
DaliyMeanConc("PM2.5", C)
}
