% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQI.R
\name{IAQI_hourly}
\alias{IAQI_hourly}
\title{Calculate hourly IAQI for given pollutant}
\usage{
IAQI_hourly(Pollu, Conc)
}
\arguments{
\item{Pollu}{name of pollutants now supports \code{SO2}, \code{NO2}, \code{CO}, and \code{O3}.}

\item{Conc}{a number of hourly mean concentration of the pollutant.}
}
\value{
the hourly IAQI index score of the give pollutant.
}
\description{
Calculate hourly IAQI for given pollutant
}
\examples{
IAQI_hourly("O3", 142)
}
